+++ frontend/src/App.tsx
@@ -6,15 +6,15 @@ import { useLocalStorage } from "react-use"
 import { User, updateUsername } from "@/utils/users"
 import { getUserApi } from "@/utils/apiUtils"
 import { UserContext } from "@/utils/context"
-import { HowTo } from "@/features/HowTo"
-import { Questions } from "@/features/Questions"
-import { Result } from "@/features/Result"
-import { Auth } from "@/features/Auth"
-import { Favorite } from "@/features/Favorite"
-import { Setting } from "@/features/Setting"
-import { NewRecipes } from "@/features/NewRecipes"
-import { NotFound } from "@/features/NotFound"
-import { supabase } from "@/features/Auth/supabaseClient"
+import { HowTo } from "@/pages/HowTo"
+import { Questions } from "@/pages/Questions"
+import { Result } from "@/pages/Result"
+import { Auth } from "@/pages/Auth"
+import { Favorite } from "@/pages/Favorite"
+import { Setting } from "@/pages/Setting"
+import { NewRecipes } from "@/pages/NewRecipes"
+import { NotFound } from "@/pages/NotFound"
+import { supabase } from "@/pages/Auth/supabaseClient"
 
 import "./App.css"
 import { Loading } from "./components/Loading"
