+++ frontend/src/pages/Auth/DeleteAccount/index.ts
