+++ frontend/src/pages/Auth/supabaseClient.ts
