+++ frontend/src/pages/Favorite/Favorite.tsx
