+++ frontend/src/pages/Favorite/index.ts
