+++ frontend/src/pages/Favorite/swiper.css
