+++ frontend/src/pages/Result/Result.tsx
