+++ frontend/src/pages/Result/index.ts
